feat(workflow): converter datasets for s3m model

+++ shybox/processing_toolkit/lib_proc_mask.py
@@ -27,7 +27,7 @@ import matplotlib.pyplot as plt
 def mask_data_by_ref(
         data: xr.DataArray,
         ref: xr.DataArray, ref_value: (float, int) = -9999.0,
-        mask_format: str = 'integer', mask_no_data: (float, int) = -9999.0, **kwargs):
+        mask_format: str = 'float', mask_no_data: (float, int) = -9999.0, **kwargs):
 
     if mask_format is not None:
         if mask_format == 'integer':
